% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loci.R
\name{get_loci}
\alias{get_loci}
\title{Get loci from GWAS file}
\usage{
get_loci(
  gwas,
  snp_col = "SNP",
  chr_col = "CHR",
  pos_col = "BP",
  maf_col = "MAF",
  beta_col = "BETA",
  se_col = "SE",
  stat_col = "NA",
  neglog10p_col = "NA",
  n_bases = 5e+05,
  p_threshold = 5e-08,
  get_ld_indep = FALSE,
  ld_pruning_r2 = 0.001,
  ld_clump_local = TRUE,
  ld_plink_bin = "plink",
  ld_bfile = "/indy/data/ld/1kg_v3/EUR"
)
}
\arguments{
\item{gwas}{A data.frame. Contains the GWAS summary statistics.}

\item{snp_col}{A string. Default="SNP". The RSID/variantID column name.}

\item{chr_col}{A string. Default="CHR". The chromosome column name.}

\item{pos_col}{A string. Default="BP". The base pair/position column name.}

\item{maf_col}{A string. Default="MAF". The MAF/minor-allele-frequency column name.}

\item{beta_col}{A string. Default="BETA". The BETA column name.}

\item{se_col}{A string. Default="SE". The SE column name.}

\item{stat_col}{A string. Default="NA". The test statistic column name. (Only required if not providing beta+se, or neglog10p)}

\item{neglog10p_col}{A string. Default="NA". The -log10 p-value column name. (Only required if not providing beta+se, or stat)}

\item{n_bases}{An interger. Default=5e5. The distance between two significant SNPs, beyond which they are defined as in separate loci.}

\item{p_threshold}{A number. Default=5e-8. P-value threshold for statistical significance}

\item{get_ld_indep}{Logical. Default=FALSE. Use Plink LD clumping to identify independent SNPs - see ieugwasr::ld_clump() docs}

\item{ld_pruning_r2}{Numeric. Default=0.001. Pruning threshold for LD.}

\item{ld_clump_local}{Logical. Default=TRUE. If clumping using local installation (rather than IEU API) - see ieugwasr::ld_clump() docs}

\item{ld_plink_bin}{A string. Default="plink". Path to Plink v1.90 binary}

\item{ld_bfile}{A string. Default="/indy/data/ld/1kg_v3/EUR". Path to BIM/BED reference panel files}
}
\value{
Returns a data frame of significant SNPs with locus annotation
}
\description{
Use distance from lead significant SNP to estimate independet loci in GWAS summary stats. Uses -log10(p) derived from BETA/SE so does not need P as input.
}
\examples{
get_loci(gwas)

}
\author{
Luke Pilling
}
