% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_gene.R
\name{get_nearest_gene}
\alias{get_nearest_gene}
\title{Get nearest gene from a set of variants}
\usage{
get_nearest_gene(
  variants,
  snp_col = "SNP",
  chr_col = "CHR",
  pos_col = "BP",
  build = 37,
  n_bases = 1e+05
)
}
\arguments{
\item{variants}{A data.frame. Contains the variants (e.g., in summary statistics).}

\item{snp_col}{A string. Default="SNP". The RSID/variantID column name.}

\item{chr_col}{A string. Default="CHR". The chromosome column name.}

\item{pos_col}{A string. Default="BP". The base pair/position column name.}

\item{build}{An integer. Default=37. Genome build to use (can only be 37 or 38).}

\item{n_bases}{An interger. Default=1e5. The max distance in base-pairs between a variant and a gene to annotate}
}
\value{
Returns a data frame of variant IDs mapped to genes (with distance)
}
\description{
Use distance variant position to get the nearest gene. Uses {snpsettest} function `map_snp_to_gene` to identify genes from GENCODE databases (https://github.com/HimesGroup/snpsettest).
}
\examples{
gwas_loci = get_loci(gwas_example)

gwas_loci_genes = get_nearest_gene(gwas_loci)

head(gwas_loci_genes)

head(gwas_loci_genes[ gwas_loci_genes$lead==TRUE , ])

}
\author{
Luke Pilling
}
