% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{lambda_gc}
\alias{lambda_gc}
\title{Estimate Lambda GC from p-values}
\usage{
lambda_gc(p)
}
\arguments{
\item{p}{a numeric vector of p-values}
}
\value{
Returns a vector
}
\description{
Estimate inflation of test statistics. Lambda GC compares the median test statistic against the expected median test statistic under the null hypothesis of no association. For well-powered quantitative traits with a known polygenic inheritance, we expect inflation of lambda GC, but for traits with no expected association, we expect lambda GC to be around 1.

Original function by Jing Hua Zhao (https://jinghuazhao.github.io/)
}
\examples{
lambda_gc(p)

}
\author{
Jing Hua Zhao
}
