% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loci.R
\name{get_loci}
\alias{get_loci}
\title{Get loci from GWAS file}
\usage{
get_loci(
  gwas,
  detect_headers = TRUE,
  snp_col = "SNP",
  chr_col = "CHR",
  pos_col = "BP",
  maf_col = "MAF",
  beta_col = "BETA",
  se_col = "SE",
  stat_col = "NA",
  p_col = "NA",
  neglog10p_col = "NA",
  use_pvalue = FALSE,
  n_bases = 5e+05,
  p_threshold = 5e-08,
  single_hla_locus = FALSE,
  hla_pos = c(2.5e+07, 3.4e+07),
  ld_clump = FALSE,
  get_ld_indep = FALSE,
  ld_clump_r2 = 0.01,
  ld_clump_local = TRUE,
  ld_plink_bin = "plink",
  ld_bfile =
    "/indy/ukbiobank/data_14631/genetics/imputed_500k/5k_eur/ukb_imp_v3.qc_sub.5k_eur",
  verbose = FALSE
)
}
\arguments{
\item{gwas}{A data.frame. Contains the GWAS summary statistics.}

\item{detect_headers}{Logical. Default=TRUE. Search input headers to see if BOLT-LMM, SAIGE, or REGENIE input (user therefore doesn't need to provide). If BOLT-LMM then automatically use p-value instead of SE.}

\item{snp_col}{A string. Default="SNP". The RSID/variantID column name.}

\item{chr_col}{A string. Default="CHR". The chromosome column name.}

\item{pos_col}{A string. Default="BP". The base pair/position column name.}

\item{maf_col}{A string. Default="MAF". The MAF/minor-allele-frequency column name.}

\item{beta_col}{A string. Default="BETA". The BETA column name.}

\item{se_col}{A string. Default="SE". The SE column name.}

\item{stat_col}{A string. Default="NA". The test statistic column name. (Only required if not providing beta+se, or neglog10p)}

\item{p_col}{A string. Default="NA". The p-value column name. (Only required if `use_pvalue==TRUE` i.e., using the provided p-value, e.g., P_BOLT_LMM)}

\item{neglog10p_col}{A string. Default="NA". The -log10 p-value column name. (Only required if not providing beta+se, or stat)}

\item{use_pvalue}{Logical. Default=FALSE. Use the provided p-value (in `p_col`) rather than computing from the test statistic? Useful for BOLT-LMM output}

\item{n_bases}{An interger. Default=5e5. The distance between two significant loci, beyond which they are defined as in separate loci.}

\item{p_threshold}{A number. Default=5e-8. P-value threshold for statistical significance}

\item{single_hla_locus}{Logical. Default=FALSE. Treat HLA as one continuous locus}

\item{hla_pos}{A numeric vector of length 2. Default=c(25e6, 34e6). The HLA region on chromosome 6 to treat as one continuous locus if `single_hla_locus==TRUE`}

\item{ld_clump}{Logical. Default=FALSE. [Previously called `get_ld_indep`] Use Plink LD clumping to identify independent SNPs - see ieugwasr::ld_clump() docs.}

\item{ld_clump_r2}{Numeric. Default=0.01. Pruning threshold for LD.}

\item{ld_clump_local}{Logical. Default=TRUE. If clumping using local installation (rather than IEU API) - see ieugwasr::ld_clump() docs}

\item{ld_plink_bin}{A string. Default="plink". Path to Plink v1.90 binary}

\item{ld_bfile}{A string. Default is to 5,000 random unrelated UK Biobank Europeans on my server :) needs a path to appropriate BIM/BED reference panel files on your server}

\item{verbose}{Logical. Default=FALSE. Be verbose}
}
\value{
Returns a data frame of significant SNPs with locus annotation
}
\description{
Use distance from lead significant SNP to estimate independet loci in GWAS summary stats. Uses -log10(p) derived from BETA/SE so does not need P as input.
}
\examples{

# distance-based loci
gwas_loci = get_loci(gwas_example)

head(gwas_loci)

head(gwas_loci[ gwas_loci$lead==TRUE , ])

# clump loci using Plink LD pruning
gwas_loci = get_loci(gwas_example, ld_clump=TRUE)

head(gwas_loci)

head(gwas_loci[ gwas_loci$lead==TRUE , ])

}
\author{
Luke Pilling
}
