% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{get_loci}
\alias{get_loci}
\title{Get loci from GWAS file}
\usage{
get_loci(
  gwas,
  snp_col = "SNP",
  chr_col = "CHR",
  pos_col = "BP",
  maf_col = "MAF",
  beta_col = "BETA",
  se_col = "SE",
  stat_col = "NA",
  neglog10p_col = "NA",
  n_bases = 1e+06,
  p_threshold = 5e-08
)
}
\arguments{
\item{gwas}{A data.frame. Contains the GWAS summary statistics.}

\item{snp_col}{A string. Default="SNP". The RSID/variantID column name.}

\item{chr_col}{A string. Default="CHR". The chromosome column name.}

\item{pos_col}{A string. Default="BP". The base pair/position column name.}

\item{maf_col}{A string. Default="MAF". The MAF/minor-allele-frequency column name.}

\item{beta_col}{A string. Default="BETA". The BETA column name.}

\item{se_col}{A string. Default="SE". The SE column name.}

\item{stat_col}{A string. Default="NA". The test statistic column name. (Only required if not providing beta+se, or neglog10p)}

\item{neglog10p_col}{A string. Default="NA". The -log10 p-value column name. (Only required if not providing beta+se, or stat)}

\item{n_bases}{An interger. Default=1e6. The distance between two significant SNPs, beyond which they are defined as in separate loci.}

\item{p_threshold}{A number. Default=5e-8. P-value threshold for statistical significance}
}
\value{
Returns a data frame of significant SNPs with locus annotation
}
\description{
Use distance from lead significant SNP to estimate independet loci in GWAS summary stats. Uses -log10(p) derived from BETA/SE so does not need P as input.
}
\examples{
get_loci(gwas)

}
\author{
Luke Pilling
}
